requirements.txt
asgiref==3.5.2
Django==4.1.4
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
